Add demo
The demo combines two modules, sending module and receiving module. 
The Sending module contains reading the raw file, combining rows to a single trip, convert it to JSON message and sending it by Kafka
The reading module contains reading message from Kafka, parsing the receiving message and save it as a dataset.